Crank-King/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
pydantic
pydantic-settings
SQLAlchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
apscheduler
